requirements.txt
opencv-python
numpy
torch
torchvision
torchaudio
ultralytics
scikit-learn
#mediapipe
Pillow
tk

